pyproject.toml
[tool.poetry]
name = "quote-analysis"
version = "0.1.0"
description = "A Django application for analyzing supplier quotes and generating historical price analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.0"
pymupdf = "^1.23.0"
langchain = "^0.1.0"
openai = "^1.0.0"
python-docx = "^1.0.0"
seaborn = "^0.13.0"
matplotlib = "^3.8.0"
pandas = "^2.2.0"
pyodbc = "^5.0.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
